Third step tutorial
- Folder name changed
- Starting of the CRUD
