+++ Ferramax/src/components/icons/IconCommunity.vue
