+++ Ferramax/src/components/icons/IconDocumentation.vue
