+++ Ferramax/src/components/icons/IconEcosystem.vue
