+++ Ferramax/src/components/icons/IconTooling.vue
