+++ Ferramax/src/shared/Client.ts
