+++ Ferramax/src/shared/Product.ts
